refactor(command): split monolithic module into focused submodules (#273)
Split lib/drivers/command/src/lib.rs (1,350 lines) into 6 focused files
following the mechanism-vs-policy principle.

New file structure:
- args.rs (109 lines) - ArgSpec, ArgKind, ArgValue
- traits.rs (135 lines) - Command, CommandHandler traits
- context.rs (255 lines) - CommandContext with accessors
- char_wait.rs (259 lines) - FindType, CharWaitOp, CharWaitContext
- result.rs (542 lines) - CommandResult and action types
- lib.rs (143 lines) - Re-exports only

Key decisions:
- Flat re-exports preserve external API unchanged
- Tests distributed across modules (47 total, zero regressions)
- Policy-adjacent types (SearchAction, UndotreeAction) documented for
  future extraction when dedicated modules exist
- FindType kept in command driver (used by CharWaitContext)

Refs #273, Parts of #150
